requirements.txt
apache-airflow==2.9.1

pandas
sqlalchemy
psycopg2-binary

apache-airflow-providers-postgres
astronomer-cosmos[dbt-postgres]
